simple sample project instead of playground for easier testing
