+++ Directions Example.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:../../Documents/Projects/Mapbox/MapboxDirections.swift/MapboxDirections.playground">
+      location = "self:Directions Example.xcodeproj">
    </FileRef>
 </Workspace>
