+++ Directions Example.xcodeproj/project.xcworkspace/xcshareddata/Directions Example.xccheckout
@@ -5,20 +5,24 @@
 	<key>IDESourceControlProjectFavoriteDictionaryKey</key>
 	<false/>
 	<key>IDESourceControlProjectIdentifier</key>
-	<string>5E8A1DAA-AAFC-44EF-8AA0-A683BD68FB89</string>
+	<string>06B36E5F-15FE-4737-BCF1-F94DDE79ABC4</string>
 	<key>IDESourceControlProjectName</key>
-	<string>MapboxDirections</string>
+	<string>project</string>
 	<key>IDESourceControlProjectOriginsDictionary</key>
 	<dict>
 		<key>0F4AF851EB4759746F2C25C1BD4E06E706014175</key>
 		<string>github.com:mapbox/mapboxdirections.swift.git</string>
+		<key>C861FC00CEE0F6A6BE81FCFF6785FAA78C58EBB3</key>
+		<string>https://github.com/SwiftyJSON/SwiftyJSON</string>
 	</dict>
 	<key>IDESourceControlProjectPath</key>
-	<string>MapboxDirections.playground</string>
+	<string>Directions Example/Directions Example.xcodeproj/project.xcworkspace</string>
 	<key>IDESourceControlProjectRelativeInstallPathDictionary</key>
 	<dict>
 		<key>0F4AF851EB4759746F2C25C1BD4E06E706014175</key>
-		<string>..</string>
+		<string>../../..</string>
+		<key>C861FC00CEE0F6A6BE81FCFF6785FAA78C58EBB3</key>
+		<string>../../..SwiftyJSON/</string>
 	</dict>
 	<key>IDESourceControlProjectURL</key>
 	<string>github.com:mapbox/mapboxdirections.swift.git</string>
@@ -35,6 +39,14 @@
 			<string>0F4AF851EB4759746F2C25C1BD4E06E706014175</string>
 			<key>IDESourceControlWCCName</key>
 			<string>MapboxDirections.swift</string>
+		</dict>
+		<dict>
+			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
+			<string>public.vcs.git</string>
+			<key>IDESourceControlWCCIdentifierKey</key>
+			<string>C861FC00CEE0F6A6BE81FCFF6785FAA78C58EBB3</string>
+			<key>IDESourceControlWCCName</key>
+			<string>SwiftyJSON</string>
 		</dict>
 	</array>
 </dict>
