+++ Directions Example/AppDelegate.swift
@@ -1,0 +1,17 @@
+import UIKit
+
+@UIApplicationMain
+
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+    var window: UIWindow?
+
+    func application(application: UIApplication, didFinishLaunchingWithOptions launchOptions: [NSObject: AnyObject]?) -> Bool {
+        window = UIWindow(frame: UIScreen.mainScreen().bounds)
+        window!.rootViewController = ViewController(nibName: nil, bundle: nil)
+        window!.makeKeyAndVisible()
+
+        return true
+    }
+
+}
