--- MapboxDirections.playground/Sources/MapboxDirections.swift
@@ -1,1 +1,0 @@
-../../MapboxDirections.swift
