--- MapboxDirections.playground/Sources/SupportCode.swift
@@ -1,3 +1,0 @@
-//
-// This file (and all other Swift source files in the Sources directory of this playground) will be precompiled into a framework which is automatically made available to MapboxDirections.playground.
-//
