--- MapboxDirections.playground/Sources/SwiftyJSON.swift
@@ -1,1 +1,0 @@
-../../SwiftyJSON/Source/SwiftyJSON.swift
