--- MapboxDirections.playground/playground.xcworkspace/xcshareddata/playground.xccheckout
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>IDESourceControlProjectFavoriteDictionaryKey</key>
-	<false/>
-	<key>IDESourceControlProjectIdentifier</key>
-	<string>D828BA8E-CA6B-44BF-987D-8CF028954412</string>
-	<key>IDESourceControlProjectName</key>
-	<string>playground</string>
-	<key>IDESourceControlProjectOriginsDictionary</key>
-	<dict>
-		<key>0F4AF851EB4759746F2C25C1BD4E06E706014175</key>
-		<string>github.com:mapbox/mapboxdirections.swift.git</string>
-	</dict>
-	<key>IDESourceControlProjectPath</key>
-	<string>MapboxDirections.playground/playground.xcworkspace</string>
-	<key>IDESourceControlProjectRelativeInstallPathDictionary</key>
-	<dict>
-		<key>0F4AF851EB4759746F2C25C1BD4E06E706014175</key>
-		<string>../..</string>
-	</dict>
-	<key>IDESourceControlProjectURL</key>
-	<string>github.com:mapbox/mapboxdirections.swift.git</string>
-	<key>IDESourceControlProjectVersion</key>
-	<integer>111</integer>
-	<key>IDESourceControlProjectWCCIdentifier</key>
-	<string>0F4AF851EB4759746F2C25C1BD4E06E706014175</string>
-	<key>IDESourceControlProjectWCConfigurations</key>
-	<array>
-		<dict>
-			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
-			<string>public.vcs.git</string>
-			<key>IDESourceControlWCCIdentifierKey</key>
-			<string>0F4AF851EB4759746F2C25C1BD4E06E706014175</string>
-			<key>IDESourceControlWCCName</key>
-			<string>MapboxDirections.swift</string>
-		</dict>
-	</array>
-</dict>
-</plist>
