--- MapboxDirections.playground/timeline.xctimeline
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Timeline
-   version = "3.0">
-   <TimelineItems>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "#CharacterRangeLen=0&amp;CharacterRangeLoc=1079&amp;EndingColumnNumber=82&amp;EndingLineNumber=23&amp;StartingColumnNumber=17&amp;StartingLineNumber=23&amp;Timestamp=451340285.769473"
-         selectedRepresentationIndex = "1"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "#CharacterRangeLen=44&amp;CharacterRangeLoc=1035&amp;EndingColumnNumber=34&amp;EndingLineNumber=22&amp;StartingColumnNumber=17&amp;StartingLineNumber=22&amp;Timestamp=451340285.769571"
-         selectedRepresentationIndex = "1"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-   </TimelineItems>
-</Timeline>
